KUDU-1933. consensus: Fix integer overflow in log index
We observed a crash on a long-running master server that looked like the
following:

  F0308 00:25:53.568773  7655 log_index.cc:171] Check failed: log_index > 0 (-2147483648 vs. 0)

It turns out that this was caused due to integer overflow on the log
index field. This patch fixes this type of truncation in a couple of
places (LogReader and MakeOpId()) and adds a couple of new tests that
fail without both of those fixes.

This patch also includes some test helper fixes to avoid log index
integer truncation in future tests.

+++ src/kudu/consensus/opid_util.cc
@@ -152,7 +152,7 @@ std::string OpsRangeString(const ConsensusRequestPB& req) {
   return ret;
 }
 
-OpId MakeOpId(int term, int index) {
+OpId MakeOpId(int64_t term, int64_t index) {
   OpId ret;
   ret.set_index(index);
   ret.set_term(term);
